+++ plugins/postcss-text-decoration-shorthand/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Text Decoration Shorthand
 
+### Unreleaed (patch)
+
+- Updated: `@csstools/color-helpers` to `v2.0.0`
+
 ### 2.2.2 (March 25, 2023)
 
 - Add `color-mix` as a known color function.